Сделать общую форму для редактирования и создания новой точки маршрута
