+++ src/presenter/main-presenter.js
@@ -4,8 +4,8 @@ import EmptyListView from '../view/empty-points-list-view';
 import TripInfoView from '../view/trip-info-view';
 import SortView from '../view/sort-view';
 import {updateItem} from '../utils/common';
-import {render, RenderPosition, replace, remove} from '../framework/render';
-import {FilterType, MODE_FORM_ADD, SORT_TYPES, SortType, UpdateType, UserAction} from '../const';
+import {remove, render, RenderPosition, replace} from '../framework/render';
+import {FilterType, SORT_TYPES, SortType, UpdateType, UserAction} from '../const';
 import {filter} from '../utils/filter';
 import {sortByDay, sortByPrice, sortByTime} from '../utils/point';
 import LoadingView from '../view/loading-view';
